backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]